refactor: domain, view 패키지를 생성하여 클래스 파일을 분배
