+++ src/main/java/baseball/domain/Player.java
@@ -1,4 +1,7 @@
-package baseball;
+package baseball.domain;
+
+import baseball.domain.Numbers;
+import baseball.view.InputView;
 
 import java.util.Scanner;
