+++ src/main/java/baseball/domain/RandomNumberGenerator.java
@@ -1,5 +1,6 @@
-package baseball;
+package baseball.domain;
 
+import utils.Constant;
 import utils.RandomUtils;
 
 public class RandomNumberGenerator {
